+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Allow audio for screen capture in Chrome and Edge browsers
 - Decouple the get call request from the UI
+- Use getSettings if possile on MediaStream
 
 ### Removed